Move functions to parse.c
